% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{simulated survival data with two stratum}
\format{
The data file consists of four components: event_time, event_status, group and strata. Data contains different event_times, two event_status (0: censored and 1: event), two groups (control and experimental) and two stratum (first and second).
}
\usage{
data(sim_data)
}
\description{
Simulated survival data for demonstrating the usage of WLRtest and SMCtest.
}
\seealso{
\code{\link{SMCtest}}, \code{\link{WLRtest}}
}
\keyword{datasets}
